requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
python-dotenv>=1.0.0
tabulate>=0.9.0
fpdf2>=2.7.0
streamlit>=1.30.0
boto3>=1.34.0
pdfplumber>=0.10.0
